Ability to QUIT currenlty running script check queues
When  4 blocks are running in parallel and a
new, smaller block shows up, we need to be able to interrupt the
script threads that are currently validating for a one of the  blocks
so that we can free up a script check queue for the new smaller  block.

Change ordering of cs_blockvalidationthread for osx and freebsd

The master was not returning because  nTodo was not being
set to zero correctly.