Block handling updates
Use only one semaphore for PV instead of two.

Move cs_blocksemaphore into globals.

Assign a scriptqueue to a running thread

Do not process a duplicate block

check whether chainwork has advanced when we initialize PV

refactor handblockmessage and put in parallel.cpp

-Add StartShutdown when closing QT Gui

StartShutdown() was never getting called which sets the shutdown
flag to true.  Generally this would be rare to cause a problem  but in PV
it can become an issue fairly easily if a user wants
to shutdown while many blocks are being connected during IBD.

-Correctly notify the UI of block tip changes

When many blocks are being connected during IBD it was
possible the the UI was not getting updated.  This is a Core
bug which happened rarely but in PV it can happen frequently.

update IsChainNearlySyncdInit after each block is processed

Rebase fixes

Run with PV when doing IBD as well as new blocks

Move cs_blockvalidationthread to parallel.h

Only print out error message when state.Invalid() or state.IsError()

When activatebestchain fails during PV due to a competing block being
beaten, we don't want to print out an error because it really isn't an
error. The block validation was terminated but the block will be store
on disk for the future in the event of a re-org.

PV.SetLocks() for SequenceLocks

remove cs_updateutxo

Do not allow two thinblocks to be validating at the same time

interconnect_nodes in parallel.py

-Change where we quit all competing threads

We must quit the competing threads before updating the UTXO
otherwise we may get a DOS from another thread when it finds
that it's inputs have been spent but before the chaintip has
had a chance to update.

- Update the UI if we are close to being syncd

Update the block Current Number of Blocks in the UI if we
are close to finishing our sync. This way when users turn off
their node for a while and turn it on they can see each block
being added as it finishes processing.

-Track whether the current block is acutally validating inputs

We need to do this in the event that two of the same block arrive
and have launched validation threads.  If this should happen then
one of those blocks will begin checking inputs before the other and
we can set a flag which indicates that fact.  Then when the other
block beginns checking inputs it can check to see if the same
block is also currently validatin and if so then exit.

-Use nChainWork to determine whether we have any work to do

When a newblock arrives we check whether the chaintip already matches
pindexMostWork.  However in PV, pindexMostWork may not necessarily point
to the chaintip, therefore we use pindexMostWork->nChainWork to determine
if an attempt should be made to connect this block.

Simplify ConnectBlock and remove unnecessary if(fParallel) statments

-mine blocks with parallel off

More Block Handling updates

-Stop all validation threads by nChainWork

We need to do this when mining our own blocks so we
only kill threads that have less work than our own.

-Don't evaluate PV.QuitReceived() unless in parallel mode

QuitReceived() gets called for each input that is checked during
connectblock() so we don't want to evaluate it unless we are really
in parallel validation.

-Re-enable SENDHEADERS when Xthins is not on

previously had to disable while debugging PV

-Keep track of Re-orgs

We need to know if a PV thread has initiated a re-org so we don't
go requesting new blocks while a re-org is in progress.

-Comment code in checkblockindex

It's not worth the problems.  If we unlock cs_main when not
in PV then it opens up a big problem with lock orders.

Missing updates for comptool from release branch

Comment out section in CheckBlockIndex()

This section does not apply to PV since blocks can possibly
still be in setBlockIndexCandidates.

Remove unused variable

Stop processing more blocks if the last one was invalid

Refactor how the threads are launched

Remove the unused pointer thread reference
Initialize the thread map after the thread has launched.